Move proxy listening to its own file
